refactor: dto -> cmd, package 구조 변경 [AIR-126]
