+++ aircnc/src/main/java/com/gurudev/aircnc/controller/RoomController.java
@@ -7,11 +7,11 @@ import static org.springframework.http.HttpStatus.CREATED;
 import com.gurudev.aircnc.configuration.jwt.JwtAuthentication;
 import com.gurudev.aircnc.controller.dto.RoomDto.RoomRegisterRequest;
 import com.gurudev.aircnc.controller.dto.RoomDto.RoomRegisterResponse;
-import com.gurudev.aircnc.domain.room.dto.RoomPhotoDto;
 import com.gurudev.aircnc.domain.room.entity.Room;
 import com.gurudev.aircnc.domain.room.entity.RoomPhoto;
 import com.gurudev.aircnc.domain.room.service.RoomPhotoService;
 import com.gurudev.aircnc.domain.room.service.RoomService;
+import com.gurudev.aircnc.domain.room.service.cmd.RoomPhotoCommand;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -43,8 +43,8 @@ public class RoomController {
         .map(roomPhotoService::upload)
         .collect(toList());
 
-    List<RoomPhotoDto> roomPhotoDtos = roomPhotos.stream()
-        .map(RoomPhotoDto::of)
+    List<RoomPhotoCommand> roomPhotoDtos = roomPhotos.stream()
+        .map(RoomPhotoCommand::of)
         .collect(toList());
 
     Room room = roomService.register(roomDto.toDto(), roomPhotoDtos, authentication.id);
