+++ aircnc/src/main/java/com/gurudev/aircnc/domain/room/service/RoomService.java
@@ -1,13 +1,13 @@
 package com.gurudev.aircnc.domain.room.service;
 
-import com.gurudev.aircnc.domain.room.dto.RoomDto;
-import com.gurudev.aircnc.domain.room.dto.RoomPhotoDto;
 import com.gurudev.aircnc.domain.room.entity.Room;
+import com.gurudev.aircnc.domain.room.service.cmd.RoomCommand;
+import com.gurudev.aircnc.domain.room.service.cmd.RoomPhotoCommand;
 import java.util.List;
 
 public interface RoomService {
 
-  Room register(RoomDto room, List<RoomPhotoDto> roomPhotos, Long hostId);
+  Room register(RoomCommand room, List<RoomPhotoCommand> roomPhotos, Long hostId);
 
   List<Room> getAll();
 }
