+++ aircnc/src/main/java/com/gurudev/aircnc/domain/room/service/RoomServiceImpl.java
@@ -2,11 +2,11 @@ package com.gurudev.aircnc.domain.room.service;
 
 import com.gurudev.aircnc.domain.member.entity.Member;
 import com.gurudev.aircnc.domain.member.repository.MemberRepository;
-import com.gurudev.aircnc.domain.room.dto.RoomDto;
-import com.gurudev.aircnc.domain.room.dto.RoomPhotoDto;
 import com.gurudev.aircnc.domain.room.entity.Room;
 import com.gurudev.aircnc.domain.room.entity.RoomPhoto;
 import com.gurudev.aircnc.domain.room.repository.RoomRepository;
+import com.gurudev.aircnc.domain.room.service.cmd.RoomCommand;
+import com.gurudev.aircnc.domain.room.service.cmd.RoomPhotoCommand;
 import com.gurudev.aircnc.exception.NotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -24,9 +24,9 @@ public class RoomServiceImpl implements RoomService {
 
   @Transactional
   @Override
-  public Room register(RoomDto roomDto, List<RoomPhotoDto> roomPhotoDtos, Long hostId) {
+  public Room register(RoomCommand roomDto, List<RoomPhotoCommand> roomPhotoDtos, Long hostId) {
     Room room = roomDto.toEntity();
-    List<RoomPhoto> roomPhotos = roomPhotoDtos.stream().map(RoomPhotoDto::toEntity).collect(
+    List<RoomPhoto> roomPhotos = roomPhotoDtos.stream().map(RoomPhotoCommand::toEntity).collect(
         Collectors.toUnmodifiableList());
 
     roomPhotos.forEach(room::addRoomPhoto);
