+++ aircnc/src/test/java/com/gurudev/aircnc/domain/util/Dto.java
@@ -1,18 +1,20 @@
 package com.gurudev.aircnc.domain.util;
 
-import com.gurudev.aircnc.domain.room.dto.RoomDto;
-import com.gurudev.aircnc.domain.room.dto.RoomPhotoDto;
+import static java.util.stream.Collectors.toList;
+
 import com.gurudev.aircnc.domain.room.entity.Address;
 import com.gurudev.aircnc.domain.room.entity.Room;
 import com.gurudev.aircnc.domain.room.entity.RoomPhoto;
+import com.gurudev.aircnc.domain.room.service.cmd.RoomCommand;
+import com.gurudev.aircnc.domain.room.service.cmd.RoomPhotoCommand;
 import java.util.List;
 
 public class Dto {
 
-  public static RoomDto of(Room room) {
+  public static RoomCommand of(Room room) {
     Address address = room.getAddress();
 
-    return new RoomDto(room.getName(),
+    return new RoomCommand(room.getName(),
         address.getLotAddress(),
         address.getRoadAddress(),
         address.getDetailedAddress(),
@@ -22,7 +24,9 @@ public class Dto {
         room.getCapacity());
   }
 
-  public static List<RoomPhotoDto> listOf(List<RoomPhoto> roomPhotos) {
-    return roomPhotos.stream().map(roomPhoto -> new RoomPhotoDto(roomPhoto.getFileName())).toList();
+  public static List<RoomPhotoCommand> listOf(List<RoomPhoto> roomPhotos) {
+    return roomPhotos.stream().map(roomPhoto -> new RoomPhotoCommand(roomPhoto.getFileName()))
+        .collect(
+            toList());
   }
 }
